# docs/apifox-dianping-openapi3.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "大众点评微服务 API 文档",
    "version": "1.0.0",
    "description": "Spring Cloud 微服务版大众点评接口文档。所有请求统一走网关 http://localhost:8080，再由网关按前缀路由到各服务（user/shop/blog/trade）。\n\n**鉴权说明**：登录成功返回 JWT token，后续请求在请求头 `authorization` 中带上该 token（直接放 token，不要加 \"Bearer \" 前缀）。未带或失效返回 HTTP 401。白名单接口（登录、商家、优惠券、热门博客、上传）无需 token。\n\n**响应统一结构**：`{success, errorMsg, data, total}`，success 为 false 时看 errorMsg。"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "网关 Gateway（生产地址按环境替换）"
    }
  ],
...
  "security": [
    { "Authorization": [] }
  ],
  "paths": {
...
    "/voucher": {
      "post": {
        "tags": ["优惠券服务"],
        "summary": "新增普通优惠券",
        "description": "普通代金券，type=0。data 返回优惠券 id。",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Voucher" },
              "example": {
                "shopId": 1,
                "title": "满100减20",
                "subTitle": "全店通用",
                "rules": "每人限购2张",
                "payValue": 20,
                "actualValue": 100,
                "type": 0
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功，data 为优惠券 id",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "errorMsg": { "type": "string", "nullable": true },
                    "data": { "type": "integer", "example": 11 },
                    "total": { "type": "integer", "nullable": true }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/voucher/seckill": {
      "post": {
        "tags": ["优惠券服务"],
        "summary": "新增秒杀券",
        "description": "type=1，需带 stock（库存，写入 tb_seckill_voucher + Redis seckill:stock:&lt;id&gt;）、beginTime、endTime。data 返回秒杀券 id。",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Voucher" },
              "example": {
                "shopId": 1,
                "title": "5折秒杀券",
                "subTitle": "限量",
                "rules": "每人限购1张",
                "payValue": 50,
                "actualValue": 100,
                "type": 1,
                "stock": 100,
                "beginTime": "2026-08-25 00:00:00",
                "endTime": "2026-08-30 23:59:59"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功，data 为秒杀券 id",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "errorMsg": { "type": "string", "nullable": true },
                    "data": { "type": "integer", "example": 12 },
                    "total": { "type": "integer", "nullable": true }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/voucher/list/{shopId}": {
      "get": {
        "tags": ["优惠券服务"],
        "summary": "查询店铺的优惠券列表",
        "description": "返回该店铺的普通券 + 秒杀券（秒杀券带 stock/beginTime/endTime）。",
        "security": [],
        "parameters": [
          {
            "name": "shopId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64", "example": 1 }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "errorMsg": { "type": "string", "nullable": true },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Voucher" }
                    },
                    "total": { "type": "integer", "nullable": true }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "Authorization": {
        "type": "apiKey",
        "in": "header",
        "name": "authorization",
        "description": "登录 /user/login 返回的 JWT token，直接放 token 值（不要加 Bearer 前缀）。"
      }
    },
    "schemas": {
...
      "Voucher": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64", "example": 12 },
          "shopId": { "type": "integer", "format": "int64", "example": 1 },
          "title": { "type": "string", "example": "5折秒杀券" },
          "subTitle": { "type": "string", "example": "限量" },
          "rules": { "type": "string", "example": "每人限购1张" },
          "payValue": { "type": "integer", "format": "int64", "description": "支付金额", "example": 50 },
          "actualValue": { "type": "integer", "format": "int64", "description": "抵扣金额", "example": 100 },
          "type": { "type": "integer", "description": "0 普通券，1 秒杀券", "example": 1 },
          "status": { "type": "integer", "description": "状态", "example": 1 },
          "stock": { "type": "integer", "description": "库存（仅秒杀券/新增秒杀券时填）", "example": 100 },
          "beginTime": { "type": "string", "format": "date-time", "description": "生效时间（秒杀券）", "example": "2026-08-25 00:00:00" },
          "endTime": { "type": "string", "format": "date-time", "description": "失效时间（秒杀券）", "example": "2026-08-30 23:59:59" },
          "createTime": { "type": "string", "format": "date-time" },
          "updateTime": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  }
}